Advance created particles from standard emitter.
When creating a particle, advance it based on the remaining time so all
particles aren't at the start. A start time is now accepted on creation,
allowing the particles to start in a non-initial state, as if time had
already passed. Update function now takes the enabled flag into account.

Added particle size parameter and random generation function. Allows for
creation of rectangular or square particles.

Particles now use a float time. While the base update function uses double
for maximum precision, this avoids constant float/double conversions when
updating particles. (where the most precise time isn't necessarily
required)

+++ modules/Particle/include/DeepSea/Particle/ParticleEmitter.h
@@ -53,7 +53,7 @@ DS_PARTICLE_EXPORT dsParticleEmitter* dsParticleEmitter_create(dsAllocator* allo
  * @param time The time that's elapsed since the last update.
  * @return False if an error occurred.
  */
-DS_PARTICLE_EXPORT bool dsParticleEmitter_update(dsParticleEmitter* emitter, double time);
+DS_PARTICLE_EXPORT bool dsParticleEmitter_update(dsParticleEmitter* emitter, float time);
 
 /**
  * @brief Destroys a particle emitter.
